Split array operator family modules
